+++ app/Http/Controllers/PostController.php
@@ -36,41 +36,6 @@ class PostController extends Controller
     }
 
     /**
-     * Like a post.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function like($post_id)
-    {
-
-        try {
-            $post_activity  = PostActivity::where('post_id', $post_id)->first();
-            assert($post_activity);
-            $post_activity->increaseLikes();
-            return $post_activity;
-        } catch (AssertionError $e) {
-            return Handler::responseWithJson($e, Handler::POST_NOT_FOUND);
-        } catch (Exception $e) {
-            return Handler::responseWithJson($e);
-        }
-    }
-    public function dislike($post_id)
-    {
-
-        try {
-            $post_activity  = PostActivity::where('post_id', $post_id)->first();
-            assert($post_activity);
-            $post_activity->decreaseLikes();
-
-            return $post_activity;
-        } catch (AssertionError $e) {
-            return Handler::responseWithJson($e, Handler::POST_NOT_FOUND);
-        } catch (Exception $e) {
-            return Handler::responseWithJson($e);
-        }
-    }
-
-    /**
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
